+++ networking/src/p2p/mod.rs
@@ -1,11 +1,8 @@
 #[macro_use]
-mod lookup;
 pub mod encoding;
-pub mod message;
-pub mod stream;
+pub mod binary_message;
+mod stream;
 pub mod peer;
-pub mod pool;
-pub mod client;
-pub mod node;
+mod pool;
 pub mod network_manager;
 pub mod network_channel;
